+++ lua_module/src/request.rs
@@ -17,3 +17,10 @@ pub fn serialize_values<T: Serialize>(
 pub fn serialize_err(err: String) -> Result<Vec<u8>, RequestError> {
     Ok(rmp_serde::to_vec(&ScriptResponse::<()>::Err(err))?)
 }
+
+/// Serializes a UnableToReachResolve error, special type to passthrough the error
+pub fn serialize_noresolve() -> Result<Vec<u8>, RequestError> {
+    Ok(rmp_serde::to_vec(
+        &ScriptResponse::<()>::UnableToReachResolve,
+    )?)
+}